Add dependency to avoid mruby 1.3.0 build error
On mruby 1.3.0, a build fails with 'fibre required for enumerator'
error. To improve this, 'mruby-fiber' should be added as a dependency
but 'mruby-enumerator' has it and it is not clear that the enumerator
which is used by other mrbgems is imported or not.

So this fixes the issue by adding 'mruby-enumerator' dependency.