Add MapOf rule
Schash supports a Hash rule but it requires specifying a key name and
doesn't allow to use an anounymous key when its value schema is only
being interested in.

This solves the issue by introducing MapOf rule like Protocol Buffer Map
which receives key and value types and checks a key type instead of a
key name existence that Hash does. As this behavior, the hash defined by
MapOf could be empty.